requirements.txt
polars-lts-cpu
awslambdaric
fsspec
s3fs
